Moved some Playwright functions into a support file.
