+++ tests/playwright/e2e/structured-expressions-selection.spec.ts
@@ -1,4 +1,5 @@
 import {Page, test, expect} from "@playwright/test";
+import {typeIndividually} from "../support/editor";
 
 let scssVars: {[varName: string]: string};
 //let strypeElIds: {[varName: string]: (...args: any[]) => string};
@@ -31,13 +32,6 @@ test.beforeEach(async ({ page }) => {
         console.log("Browser log:", msg.text());
     });
 });
-
-async function typeIndividually(page: Page, content: string) {
-    for (let i = 0; i < content.length; i++) {
-        await page.keyboard.type(content[i]);
-        await page.waitForTimeout(75);
-    }
-}
 
 async function getSelection(page: Page) : Promise<{ id: string, cursorPos : number }> {
     // We need a delay to make sure last DOM update has occurred:
